+++ project/newsfeed/models.py
@@ -74,27 +74,3 @@ class Comment(NewsfeedObject):
 
     def get_user_url(self):
         return f"/api/v1/user/{self.author}/"
-
-
-class Notice(models.Model):
-
-    id = models.AutoField(primary_key=True)
-    user = models.ForeignKey(User, on_delete=CASCADE, related_name="notices")
-    post = models.ForeignKey(Post, on_delete=CASCADE, null=True, related_name="notices")
-    parent_comment = models.ForeignKey(
-        Comment, on_delete=CASCADE, null=True, related_name="notices"
-    )
-    content = models.CharField(max_length=30, blank=True)
-    created = models.DateTimeField(auto_now=True)
-    is_checked = models.BooleanField(default=False)
-    url = models.CharField(max_length=1000)
-
-
-class NoticeSender(models.Model):
-    notice = models.ForeignKey(
-        Notice, on_delete=CASCADE, null=True, related_name="senders"
-    )
-    user = models.ForeignKey(
-        User, on_delete=CASCADE, null=True, related_name="noticesenders"
-    )
-    count = models.PositiveIntegerField(default=0)
