+++ project/newsfeed/urls.py
@@ -1,19 +1,14 @@
-from django.urls import path, include
+from django.urls import path
 from .views import (
     PostListView,
     PostUpdateView,
     PostLikeView,
     CommentListView,
     CommentLikeView,
-    NoticeView,
-    NoticeListView,
     CommentUpdateDeleteView,
-    NoticeOnOffView,
-    TestView
+    TestView,
 )
-from rest_framework.routers import SimpleRouter
-from django.conf import settings
-from django.conf.urls.static import static
+from notice.views import NoticeOnOffView
 
 urlpatterns = [
     path("newsfeed/", PostListView.as_view()),
@@ -23,7 +18,5 @@ urlpatterns = [
     path("newsfeed/<int:post_id>/<int:comment_id>/", CommentUpdateDeleteView.as_view()),
     path("newsfeed/<int:post_id>/like/", PostLikeView.as_view()),
     path("newsfeed/<int:post_id>/<int:comment_id>/like/", CommentLikeView.as_view()),
-    path("newsfeed/notices/", NoticeListView.as_view()),
-    path("newsfeed/notices/<int:notice_id>/", NoticeView.as_view()),
-    path("test/", TestView.as_view())
+    path("test/", TestView.as_view()),
 ]
